[US_PA] Hydrate state_staff_supervisor_periods (Recidiviz/recidiviz-data#19870)
## Description of the change

This hydrates the StateStaffSupervisorPeriod entity so that we can
deprecate the `us_pa_agent_supervisor_preprocessed` view.

The view to hydrate this entity uses the same data source as the
preprocessed view (`dbo_RelAgentHistory`). Each row represents a period
during which a parole officer had a particular supervisor. Periods where
a parole officer did not have a supervisor (usually a week or less) are
assumed to be a data entry anomaly and excluded; the period of no
supervisor is combined with the period of the previous supervisor.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d5e123f972922eeb9b6fb57c2b21b5b9973fa143