Adds supervision unit-level metrics to aggregated metrics pipeline (Recidiviz/recidiviz-data#19667)
## Description of the change

1) Joins in `state_staff_location_period` + `location_metadata` to
create a sessionized table of staff attributes called
`supervision_officer_attributes_sessions`.
2) Creates `supervision_unit_sessions` by joining
`supervision_officer_sessions` with
`supervision_officer_attribute_sessions` to get continuous periods of
client assignment to a supervision unit via their officer.
2) Adds `SUPERVISION_UNIT` as a new aggregation level type and adds this
to the supported levels in the aggregated metrics pipeline. This
includes adding `SUPERVISION_UNIT` to the list of supported levels in
`misc_aggregated_metrics`, and enables calculation of misc metrics
currently supporting district/office/state level to units as well.

Also renames what was previously `primary_office`/`primary_district` to
be instead called `supervision_office_inferred` and
`supervision_district_inferred`. Roster-based locations are called
`supervision_office`/`supervision_district`/`supervision_unit` etc., and
are joined in on the last day of every analysis period (this means that
if an officer is in Unit 2 for the first 20 days of the month, and Unit
5 for the last 10 days of the month, we will associate Unit 5 with the
officer for that month).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#19603

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: 4ce1a950396b4506275a996a2e4807cd54f10695