[Cleanup] Prepare to delete the EnumOverrides class (Recidiviz/recidiviz-data#20047)
## Description of the change

Updated the `EnumMappingManifest` so that it stores a `StrictEnumParser`
directly instead of an `EnumOverrides` object. In order to make this
change, updated the `StrictEnumParser` to build `EnumOverrides`
internally.

With this change, `StrictEnumParser` is the ONLY remaining place that
references `EnumOverrides`. In a subsequent change, I will eliminate
`EnumOverrides` entirely and roll the necessary functionality into
`StrictEnumParser`.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#8905.

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: c84dd58ac554352b5787e17fc33adde66981e7a4