[Normalization] Clearing release_reason_raw_text for erroneous future periods (Recidiviz/recidiviz-data#20121)
## Description of the change

[This
validation](https://app.zenhub.com/workspaces/all-pods-6324d1d66be1c200119ab65a/issues/gh/recidiviz/recidiviz-data/20107)
started failing once we added in segregation data to incarceration
periods in TN because some of the segregation data includes inferred
periods of segregation in the future (Meaning they will estimate an end
date for a given person to end their segregation stay that is in the
future.)

Dealing with excluding these future dates in raw data is something we
can handle in the view itself, but the actual validation is failing
because in normalization, we correct periods with erroneous future dates
by clearing the date and release_reason, but we don't currently clear
the release_reason_raw_text. When we are correcting these periods in
normalization, we should wipe both release_reason and
release_reason_raw_text since we are clearing all estimated future
information from the period.

This PR adds in a line to the delegate that overwrites the IP raw_text
field to None.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#20107

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [X] This pull request has a descriptive title and information useful
to a reviewer
- [X] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [X] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: b26575217e04ad0b8d0e7be734ca8e12a0a75906