[US_MI] Add normalization logic for projected_max_release_date (Recidiviz/recidiviz-data#20116)
## Description of the change

Adding some logic to normalization that calculates
`projected_max_release_date` as `effective_date` + `max_length_days` in
cases where `projected_max_release_date` is originally null (which
happens when the sentence record in the raw table
`ADH_OFFENDER_SENTENCE` has a null `expiration_date`).

Ideally, we'd like to have this in the mapping directly but I'm putting
it in normalization for now so that we can use this revision sooner and
will move this logic to the mapping when I next do a full MI rerun.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7763a24113610d628b3386b3325af007f9569757