Deprecate unused revocation and reincarceration sessions views (Recidiviz/recidiviz-data#20150)
## Description of the change
This PR makes two changes:
1. Deprecates sessions and validation views related to revocation and
reincarceration sessions that are not being used downstream
2. Updates the supervision super session definition to include
WEEKEND_CONFINEMENT

The deprecated views are the following:
1. `revocation_cohort_sessions` - this is a view that unnests revocation
sessions by a month index and is could be used to create cumulative
revocation curves. This use is now replaced by either the cohort event
dashboard or the cohort event functions in Python
2. `reincarceration_cohort_sessions` - same as above but for
reincarcerations
3. `reincarceration_sessions_from_dataflow` - this is a view that
identifies reincarcerations using the dataflow metric
`most_recent_recidivism_count_metrics_materialized`. This view is not
being used downstream.
4. `revocation_sessions_to_dataflow_disaggregated` - this is a
validation view that was created to compare revocation sessions to the
the commitment from supervision dataflow metric. It is not a configured
validation and is not being used downstream
5. `reincarcerations_from_dataflow_to dataflow_disaggregated` - similar
to above but a comparison of reincarceration sessions with with the
recidivism count metric
6. `reincarcerations from sessions_to_dataflow_disaggregated` - same as
above

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 5fc221293acd246f6f6865069c9fb99456e97bac