[US_NC][Not-yet-partner] Hydrate StateIncarcerationPeriod (Recidiviz/recidiviz-data#20141)
## Description of the change

Extra notes included in the view and mappings files.

This hydrates StateIncarcerationPeriod information based on sentence
information found in raw data table `INMT4BB1` ("Sentence
Computations"). Some of these sentences have started and ended, and
others are currently active. Sentences are joined with raw data table
`INMT4AA1` ("Inmate Profile") which includes details about _only the
most recent_ sentence a person has received, and the _single most recent
movement_ that person has made. So, if a person's last movement in the
correctional system happened on the date that their sentence began or
ended, the nature of that movement is stored as their `admission_reason`
or `release_reason`.

There are only a few cases where the results of this join to hydrate
`admission_reason` and `release_reason` don't make logical sense; they
are noted in the mapping .yaml, and they seem to result largely from
individual differences in how movement type codes are used.

The one exception seems to happen in the case when someone was released
to supervision before the end date of their sentence, and then _more
recently_ were released from parole on the date their sentence ended. In
this case, the "most recent movement type" is `TERMINATED PAROLE` rather
than `RELEASE TO PAROLE`. This issue would be resolved if we had access
to _all_ of a person's movements rather than only their most recent one.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 65ce90a5e514fc3d05210ac010b2fe74449b31a9