[Admin Panel] Changing casing for page names with Population in the title (Recidiviz/recidiviz-data#20123)
## Description of the change

Related to Recidiviz/recidiviz-data#19394, This was a fast hack and I'm not actually sure if
it's the right way to do it since this is very outside my wheelhosue,
but the `.replace(/Po/g, "PO")` line captures (in addition to correctly
all instances of "PO") all instances of `Population` and changes it in
the tab view to `POpulation`. Assuming ordering matters, this fix would
make sure any time the line above it changed Population to POpulation,
it is corrected back to correct camel casing. Also added in "PFI" since
that's used in validations.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#19729

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [X] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [X] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d1983ed8793b0f52fc4108e22847b8038793f0d6