[Justice Counts] [Bug] save automatic upload id as -1 (Recidiviz/recidiviz-data#20215)
## Description of the change

This change

1) makes the `DatapointHistory` `user_account_id` field nullable. This
field cannot hold the `AUTOMATIC_UPLOAD_ID` constant because it has to
be tied to a real account in the `UserAccount` table.

2) Saves automatic upload changed as `AUTOMATIC_UPLOAD_ID` in the report
`modified_by` column because that column holds an array that must only
hold numbers.

Alternatives considered : create a user account with an id that is
negative 1 in staging and prod.

## Related issues

Closes Recidiviz/recidiviz-data#20213

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 09feaec153339940b0be73e12757813b44e424bf