[US_MO] Amend fix to overlapping supervision periods view to use correct CTEs (Recidiviz/recidiviz-data#20286)
## Description of the change

Recidiviz/recidiviz-data#19282 changed the order of CTEs in the supervision periods view to
ensure that overlapping field assignment periods get dropped properly.
However, the names referenced in a later CTE was not changed, meaning
that the overlapping periods weren't actually getting dropped. This
change fixes the reference to the wrong CTE.

Also included in this change is a documentation update for the unrelated
table OFNDR_PDB_OFNDR_ASMNTS, which had its config edited in Recidiviz/recidiviz-data#20191 but
the documentation failed to automatically update accordingly.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] This pull request has been moved out of a Draft state, has no
"Work In Progress" label, and has assigned reviewers
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: cd068b02111a0caaedb5725773b548bd51e1f930